Added more consistent 'log_message' calls to the hardware sid drivers.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@31299 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/unix/ps.h
@@ -29,9 +29,9 @@
 
 #include "types.h"
 
-extern void ps_file_out_ctr(WORD parsid_ctrport, int chipno);
-extern void ps_io_out_ctr(WORD parsid_ctrport, int chipno);
-extern void ps_ieee1284_out_ctr(WORD parsid_ctrport, int chipno);
+extern void ps_file_out_ctr(BYTE parsid_ctrport, int chipno);
+extern void ps_io_out_ctr(BYTE parsid_ctrport, int chipno);
+extern void ps_ieee1284_out_ctr(BYTE parsid_ctrport, int chipno);
 
 extern BYTE ps_file_in_ctr(int chipno);
 extern BYTE ps_io_in_ctr(int chipno);
